alx_travel_app/requirement.txt
asgiref>=3.8.1
python-dotenv
Django==4.2.25
djangorestframework
django-cors-headers
drf-yasg
django-environ
celery
psycopg2
requests